+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Enhancements
 * `FuncArgs` is also implemented for arrays.
 * `Engine::set_XXX` API can now be chained.
 * `EvalContext::scope_mut` now returns `&mut Scope` instead of `&mut &mut Scope`.
+* Line-style doc-comments are now merged into a single string to avoid creating many strings. Block-style doc-comments continue to be independent strings.
 
 
 Version 1.11.0